+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+* Fix Content-Length parsing
+
 Version 287:
 
 * Remove CODE_OF_CONDUCT.md